add additional python classes
