+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Moved classes to external source (outside of main)
 - Token authentication
 - More actions than `GET`/`DELETE`
