+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
 
 setuptools.setup(
     name="restify-ENGYAK",  # Replace with your own username
-    version="0.1.2",
+    version="0.1.3",
     author="Nicholas Schmidt",
     author_email="nick.schmidt3@gmail.com",
     description="Abstractions should save typing, not thinking! This project will provide a quick and simple CLI for REST API consumption. Build your own library of known API endpoints/modifiers, and consume directly via the CLI.",
